+++ src/main/java/com/googongill/aditory/repository/CategoryLikeRepository.java
@@ -12,7 +12,7 @@ public interface CategoryLikeRepository extends JpaRepository<CategoryLike, Long
 
     boolean existsByUserAndCategory(User user, Category category);
 
-    Long countByCategory(Category category);
+    Integer countByCategory(Category category);
 
     Optional<CategoryLike> findByUserAndCategory(User user, Category category);
 }
